added account credit has one model and notice menu tolltip & notification menu to header and added new span notices

+++ db/migrate/20111110112317_create_service_prices.rb
@@ -2,8 +2,8 @@ class CreateServicePrices < ActiveRecord::Migration
   def change
     create_table :service_prices do |t|
       t.references :service
-      t.float :sender_credit, :default => 0
-      t.float :receiver_credit, :default => 0 
+      t.integer :sender_credit, :default => 0
+      t.integer :receiver_credit, :default => 0 
 
       t.timestamps
     end
